organizing better
