+++ head node/includes/node_status.h
@@ -1,13 +1,9 @@
 #ifndef node_status_h_INCLUDED
 #define node_status_h_INCLUDED
 
-typedef struct {
-	float cpuUse;
-	int num_cores;
-	float coreUse[12];
-	float ram_usage;
-} NODEINFO;
+#include "types.h"
 
 NODEINFO * node_info(char * hostname);
+void updateNodeHealth(void);
 
 #endif // node_status_h_INCLUDED
