+++ head node/includes/nodes_list.h
@@ -1,7 +1,7 @@
 #ifndef gen_nodes_list_h_INCLUDED
 #define gen_nodes_list_h_INCLUDED
 
-#include "cJSON.h"
+#include "../src/cJSON.h"
 
 cJSON * check_nodes(const char * ip);
 void update_status(int core, char * host);
